Crate rusoto_discovery

AWS Application Discovery Service
AWS Application Discovery Service helps you plan application migration projects by automatically identifying servers, virtual machines (VMs), software, and software dependencies running in your on-premises data centers. Application Discovery Service also collects application performance data, which can help you assess the outcome of your migration. The data collected by Application Discovery Service is securely retained in an AWS-hosted and managed database in the cloud. You can export the data as a CSV or XML file into your preferred visualization tool or cloud-migration solution to plan your migration. For more information, see AWS Application Discovery Service FAQ.

Application Discovery Service offers two modes of operation:

Agentless discovery mode is recommended for environments that use VMware vCenter Server. This mode doesn't require you to install an agent on each host. Agentless discovery gathers server information regardless of the operating systems, which minimizes the time required for initial on-premises infrastructure assessment. Agentless discovery doesn't collect information about software and software dependencies. It also doesn't work in non-VMware environments.

Agent-based discovery mode collects a richer set of data than agentless discovery by using the AWS Application Discovery Agent, which you install on one or more hosts in your data center. The agent captures infrastructure and application information, including an inventory of installed software applications, system and process performance, resource utilization, and network dependencies between workloads. The information collected by agents is secured at rest and in transit to the Application Discovery Service database in the cloud.

We recommend that you use agent-based discovery for non-VMware environments and to collect information about software and software dependencies. You can also run agent-based and agentless discovery simultaneously. Use agentless discovery to quickly complete the initial infrastructure assessment and then install agents on select hosts.

Application Discovery Service integrates with application discovery solutions from AWS Partner Network (APN) partners. Third-party application discovery tools can query Application Discovery Service and write to the Application Discovery Service database using a public API. You can then import the data into either a visualization tool or cloud-migration solution.

Application Discovery Service doesn't gather sensitive information. All data is handled according to the AWS Privacy Policy. You can operate Application Discovery Service offline to inspect collected data before it is shared with the service.

This API reference provides descriptions, syntax, and usage examples for each of the actions and data types for Application Discovery Service. The topic for each action shows the API request parameters and the response. Alternatively, you can use one of the AWS SDKs to access an API that is tailored to the programming language or platform that you're using. For more information, see AWS SDKs.

This guide is intended for use with the AWS Application Discovery Service User Guide .

If you're using the service, you're probably looking for DiscoveryClient and Discovery.

Structs

AgentConfigurationStatus
Information about agents or connectors that were instructed to start collecting data. Information includes the agent/connector ID, a description of the operation, and whether the agent/connector configuration was updated.

AgentInfo
Information about agents or connectors associated with the user’s AWS account. Information includes agent or connector IDs, IP addresses, media access control (MAC) addresses, agent or connector health, hostname where the agent or connector resides, and agent version for each agent.

AgentNetworkInfo
Network details about the host where the agent/connector resides.

AssociateConfigurationItemsToApplicationRequest

AssociateConfigurationItemsToApplicationResponse

BatchDeleteImportDataRequest

BatchDeleteImportDataResponse

ConfigurationTag
Tags for a configuration item. Tags are metadata that help you categorize IT assets.

ContinuousExportDescription
A list of continuous export descriptions.

CreateApplicationRequest

CreateApplicationResponse

CreateTagsRequest

CreateTagsResponse

CustomerAgentInfo
Inventory data for installed discovery agents.

CustomerConnectorInfo
Inventory data for installed discovery connectors.

DeleteApplicationsRequest

DeleteApplicationsResponse

DeleteTagsRequest

DeleteTagsResponse

DescribeAgentsRequest

DescribeAgentsResponse

DescribeConfigurationsRequest

DescribeConfigurationsResponse

DescribeContinuousExportsRequest

DescribeContinuousExportsResponse

DescribeExportConfigurationsRequest

DescribeExportConfigurationsResponse

DescribeExportTasksRequest

DescribeExportTasksResponse

DescribeImportTasksRequest

DescribeImportTasksResponse

DescribeTagsRequest

DescribeTagsResponse

DisassociateConfigurationItemsFromApplicationRequest

DisassociateConfigurationItemsFromApplicationResponse

DiscoveryBatchDeleteImportDataError
Error messages returned for each import task that you deleted as a response for this command.

DiscoveryClient
A client for the AWS Application Discovery Service API.

ExportConfigurationsResponse

ExportFilter
Used to select which agent's data is to be exported. A single agent ID may be selected for export using the StartExportTask action.

ExportInfo
Information regarding the export status of discovered data. The value is an array of objects.

Filter
A filter that can use conditional operators.

For more information about filters, see Querying Discovered Configuration Items.

GetDiscoverySummaryRequest

GetDiscoverySummaryResponse

ImportTask
An array of information related to the import task request that includes status information, times, IDs, the Amazon S3 Object URL for the import file, and more.

ImportTaskFilter

A name-values pair of elements you can use to filter the results when querying your import tasks. Currently, wildcards are not supported for filters.

When filtering by import status, all other filter values are ignored.

ListConfigurationsRequest

ListConfigurationsResponse

ListServerNeighborsRequest

ListServerNeighborsResponse

NeighborConnectionDetail
Details about neighboring servers.

OrderByElement
A field and direction for ordered output.

StartContinuousExportRequest

StartContinuousExportResponse

StartDataCollectionByAgentIdsRequest

StartDataCollectionByAgentIdsResponse

StartExportTaskRequest

StartExportTaskResponse

StartImportTaskRequest

StartImportTaskResponse

StopContinuousExportRequest

StopContinuousExportResponse

StopDataCollectionByAgentIdsRequest

StopDataCollectionByAgentIdsResponse

Tag
Metadata that help you categorize IT assets.

TagFilter
The tag filter. Valid names are: tagKey, tagValue, configurationId.

UpdateApplicationRequest

UpdateApplicationResponse

Enums

AssociateConfigurationItemsToApplicationError
Errors returned by AssociateConfigurationItemsToApplication

BatchDeleteImportDataError
Errors returned by BatchDeleteImportData

CreateApplicationError
Errors returned by CreateApplication

CreateTagsError
Errors returned by CreateTags

DeleteApplicationsError
Errors returned by DeleteApplications

DeleteTagsError
Errors returned by DeleteTags

DescribeAgentsError
Errors returned by DescribeAgents

DescribeConfigurationsError
Errors returned by DescribeConfigurations

DescribeContinuousExportsError
Errors returned by DescribeContinuousExports

DescribeExportConfigurationsError
Errors returned by DescribeExportConfigurations

DescribeExportTasksError
Errors returned by DescribeExportTasks

DescribeImportTasksError
Errors returned by DescribeImportTasks

DescribeTagsError
Errors returned by DescribeTags

DisassociateConfigurationItemsFromApplicationError
Errors returned by DisassociateConfigurationItemsFromApplication

ExportConfigurationsError
Errors returned by ExportConfigurations

GetDiscoverySummaryError
Errors returned by GetDiscoverySummary

ListConfigurationsError
Errors returned by ListConfigurations

ListServerNeighborsError
Errors returned by ListServerNeighbors

StartContinuousExportError
Errors returned by StartContinuousExport

StartDataCollectionByAgentIdsError
Errors returned by StartDataCollectionByAgentIds

StartExportTaskError
Errors returned by StartExportTask

StartImportTaskError
Errors returned by StartImportTask

StopContinuousExportError
Errors returned by StopContinuousExport

StopDataCollectionByAgentIdsError
Errors returned by StopDataCollectionByAgentIds

UpdateApplicationError
Errors returned by UpdateApplication

Traits

Discovery
Trait representing the capabilities of the AWS Application Discovery Service API. AWS Application Discovery Service clients implement this trait.