Trait rusoto_discovery::Discovery

pub trait Discovery {
    fn associate_configuration_items_to_application(
        &self, 
        input: AssociateConfigurationItemsToApplicationRequest
    ) -> RusotoFuture<AssociateConfigurationItemsToApplicationResponse, AssociateConfigurationItemsToApplicationError>;

    fn batch_delete_import_data(
        &self, 
        input: BatchDeleteImportDataRequest
    ) -> RusotoFuture<BatchDeleteImportDataResponse, BatchDeleteImportDataError>;

    fn create_application(
        &self, 
        input: CreateApplicationRequest
    ) -> RusotoFuture<CreateApplicationResponse, CreateApplicationError>;

    fn create_tags(
        &self, 
        input: CreateTagsRequest
    ) -> RusotoFuture<CreateTagsResponse, CreateTagsError>;

    fn delete_applications(
        &self, 
        input: DeleteApplicationsRequest
    ) -> RusotoFuture<DeleteApplicationsResponse, DeleteApplicationsError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsRequest
    ) -> RusotoFuture<DeleteTagsResponse, DeleteTagsError>;

    fn describe_agents(
        &self, 
        input: DescribeAgentsRequest
    ) -> RusotoFuture<DescribeAgentsResponse, DescribeAgentsError>;

    fn describe_configurations(
        &self, 
        input: DescribeConfigurationsRequest
    ) -> RusotoFuture<DescribeConfigurationsResponse, DescribeConfigurationsError>;

    fn describe_continuous_exports(
        &self, 
        input: DescribeContinuousExportsRequest
    ) -> RusotoFuture<DescribeContinuousExportsResponse, DescribeContinuousExportsError>;

    fn describe_export_configurations(
        &self, 
        input: DescribeExportConfigurationsRequest
    ) -> RusotoFuture<DescribeExportConfigurationsResponse, DescribeExportConfigurationsError>;

    fn describe_export_tasks(
        &self, 
        input: DescribeExportTasksRequest
    ) -> RusotoFuture<DescribeExportTasksResponse, DescribeExportTasksError>;

    fn describe_import_tasks(
        &self, 
        input: DescribeImportTasksRequest
    ) -> RusotoFuture<DescribeImportTasksResponse, DescribeImportTasksError>;

    fn describe_tags(
        &self, 
        input: DescribeTagsRequest
    ) -> RusotoFuture<DescribeTagsResponse, DescribeTagsError>;

    fn disassociate_configuration_items_from_application(
        &self, 
        input: DisassociateConfigurationItemsFromApplicationRequest
    ) -> RusotoFuture<DisassociateConfigurationItemsFromApplicationResponse, DisassociateConfigurationItemsFromApplicationError>;

    fn export_configurations(
        &self
    ) -> RusotoFuture<ExportConfigurationsResponse, ExportConfigurationsError>;

    fn get_discovery_summary(
        &self
    ) -> RusotoFuture<GetDiscoverySummaryResponse, GetDiscoverySummaryError>;

    fn list_configurations(
        &self, 
        input: ListConfigurationsRequest
    ) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>;

    fn list_server_neighbors(
        &self, 
        input: ListServerNeighborsRequest
    ) -> RusotoFuture<ListServerNeighborsResponse, ListServerNeighborsError>;

    fn start_continuous_export(
        &self
    ) -> RusotoFuture<StartContinuousExportResponse, StartContinuousExportError>;

    fn start_data_collection_by_agent_ids(
        &self, 
        input: StartDataCollectionByAgentIdsRequest
    ) -> RusotoFuture<StartDataCollectionByAgentIdsResponse, StartDataCollectionByAgentIdsError>;

    fn start_export_task(
        &self, 
        input: StartExportTaskRequest
    ) -> RusotoFuture<StartExportTaskResponse, StartExportTaskError>;

    fn start_import_task(
        &self, 
        input: StartImportTaskRequest
    ) -> RusotoFuture<StartImportTaskResponse, StartImportTaskError>;

    fn stop_continuous_export(
        &self, 
        input: StopContinuousExportRequest
    ) -> RusotoFuture<StopContinuousExportResponse, StopContinuousExportError>;

    fn stop_data_collection_by_agent_ids(
        &self, 
        input: StopDataCollectionByAgentIdsRequest
    ) -> RusotoFuture<StopDataCollectionByAgentIdsResponse, StopDataCollectionByAgentIdsError>;

    fn update_application(
        &self, 
        input: UpdateApplicationRequest
    ) -> RusotoFuture<UpdateApplicationResponse, UpdateApplicationError>;
}

Trait representing the capabilities of the AWS Application Discovery Service API. AWS Application Discovery Service clients implement this trait.

Required methods

fn associate_configuration_items_to_application(
&self,
input: AssociateConfigurationItemsToApplicationRequest
) -> RusotoFuture<AssociateConfigurationItemsToApplicationResponse, AssociateConfigurationItemsToApplicationError>

Associates one or more configuration items with an application.

fn batch_delete_import_data(
&self,
input: BatchDeleteImportDataRequest
) -> RusotoFuture<BatchDeleteImportDataResponse, BatchDeleteImportDataError>

Deletes one or more import tasks, each identified by their import ID. Each import task has a number of records that can identify servers or applications.

AWS Application Discovery Service has built-in matching logic that will identify when discovered servers match existing entries that you've previously discovered, the information for the already-existing discovered server is updated. When you delete an import task that contains records that were used to match, the information in those matched records that comes from the deleted records will also be deleted.

fn create_application(
&self,
input: CreateApplicationRequest
) -> RusotoFuture<CreateApplicationResponse, CreateApplicationError>

Creates an application with the given name and description.

fn create_tags(
&self,
input: CreateTagsRequest
) -> RusotoFuture<CreateTagsResponse, CreateTagsError>

Creates one or more tags for configuration items. Tags are metadata that help you categorize IT assets. This API accepts a list of multiple configuration items.

fn delete_applications(
&self,
input: DeleteApplicationsRequest
) -> RusotoFuture<DeleteApplicationsResponse, DeleteApplicationsError>

Deletes a list of applications and their associations with configuration items.

fn delete_tags(
&self,
input: DeleteTagsRequest
) -> RusotoFuture<DeleteTagsResponse, DeleteTagsError>

Deletes the association between configuration items and one or more tags. This API accepts a list of multiple configuration items.

fn describe_agents(
&self,
input: DescribeAgentsRequest
) -> RusotoFuture<DescribeAgentsResponse, DescribeAgentsError>

Lists agents or connectors as specified by ID or other filters. All agents/connectors associated with your user account can be listed if you call DescribeAgents as is without passing any parameters.

fn describe_configurations(
&self,
input: DescribeConfigurationsRequest
) -> RusotoFuture<DescribeConfigurationsResponse, DescribeConfigurationsError>

Retrieves attributes for a list of configuration item IDs.

All of the supplied IDs must be for the same asset type from one of the following:

server

application

process

connection

Output fields are specific to the asset type specified. For example, the output for a server configuration item includes a list of attributes about the server, such as host name, operating system, number of network cards, etc.

For a complete list of outputs for each asset type, see Using the DescribeConfigurations Action.

fn describe_continuous_exports(
&self,
input: DescribeContinuousExportsRequest
) -> RusotoFuture<DescribeContinuousExportsResponse, DescribeContinuousExportsError>

Lists exports as specified by ID. All continuous exports associated with your user account can be listed if you call DescribeContinuousExports as is without passing any parameters.

fn describe_export_configurations(
&self,
input: DescribeExportConfigurationsRequest
) -> RusotoFuture<DescribeExportConfigurationsResponse, DescribeExportConfigurationsError>

DescribeExportConfigurations is deprecated. Use DescribeImportTasks, instead.

fn describe_export_tasks(
&self,
input: DescribeExportTasksRequest
) -> RusotoFuture<DescribeExportTasksResponse, DescribeExportTasksError>

Retrieve status of one or more export tasks. You can retrieve the status of up to 100 export tasks.

fn describe_import_tasks(
&self,
input: DescribeImportTasksRequest
) -> RusotoFuture<DescribeImportTasksResponse, DescribeImportTasksError>

Returns an array of import tasks for your account, including status information, times, IDs, the Amazon S3 Object URL for the import file, and more.

fn describe_tags(
&self,
input: DescribeTagsRequest
) -> RusotoFuture<DescribeTagsResponse, DescribeTagsError>

Retrieves a list of configuration items that have tags as specified by the key-value pairs, name and value, passed to the optional parameter filters.

There are three valid tag filter names:

tagKey

tagValue

configurationId

Also, all configuration items associated with your user account that have tags can be listed if you call DescribeTags as is without passing any parameters.

fn disassociate_configuration_items_from_application(
&self,
input: DisassociateConfigurationItemsFromApplicationRequest
) -> RusotoFuture<DisassociateConfigurationItemsFromApplicationResponse, DisassociateConfigurationItemsFromApplicationError>

Disassociates one or more configuration items from an application.

fn export_configurations(
&self
) -> RusotoFuture<ExportConfigurationsResponse, ExportConfigurationsError>

Deprecated. Use StartExportTask instead.

Exports all discovered configuration data to an Amazon S3 bucket or an application that enables you to view and evaluate the data. Data includes tags and tag associations, processes, connections, servers, and system performance. This API returns an export ID that you can query using the DescribeExportConfigurations API. The system imposes a limit of two configuration exports in six hours.

fn get_discovery_summary(
&self
) -> RusotoFuture<GetDiscoverySummaryResponse, GetDiscoverySummaryError>

Retrieves a short summary of discovered assets.

This API operation takes no request parameters and is called as is at the command prompt as shown in the example.

fn list_configurations(
&self,
input: ListConfigurationsRequest
) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>

Retrieves a list of configuration items as specified by the value passed to the required paramater configurationType. Optional filtering may be applied to refine search results.

fn list_server_neighbors(
&self,
input: ListServerNeighborsRequest
) -> RusotoFuture<ListServerNeighborsResponse, ListServerNeighborsError>

Retrieves a list of servers that are one network hop away from a specified server.

fn start_continuous_export(
&self
) -> RusotoFuture<StartContinuousExportResponse, StartContinuousExportError>

Start the continuous flow of agent's discovered data into Amazon Athena.

fn start_data_collection_by_agent_ids(
&self,
input: StartDataCollectionByAgentIdsRequest
) -> RusotoFuture<StartDataCollectionByAgentIdsResponse, StartDataCollectionByAgentIdsError>

Instructs the specified agents or connectors to start collecting data.

fn start_export_task(
&self,
input: StartExportTaskRequest
) -> RusotoFuture<StartExportTaskResponse, StartExportTaskError>

Begins the export of discovered data to an S3 bucket.

If you specify agentIds in a filter, the task exports up to 72 hours of detailed data collected by the identified Application Discovery Agent, including network, process, and performance details. A time range for exported agent data may be set by using startTime and endTime. Export of detailed agent data is limited to five concurrently running exports.

If you do not include an agentIds filter, summary data is exported that includes both AWS Agentless Discovery Connector data and summary data from AWS Discovery Agents. Export of summary data is limited to two exports per day.

fn start_import_task(
&self,
input: StartImportTaskRequest
) -> RusotoFuture<StartImportTaskResponse, StartImportTaskError>

Starts an import task, which allows you to import details of your on-premises environment directly into AWS without having to use the Application Discovery Service (ADS) tools such as the Discovery Connector or Discovery Agent. This gives you the option to perform migration assessment and planning directly from your imported data, including the ability to group your devices as applications and track their migration status.

To start an import request, do this:

Download the specially formatted comma separated value (CSV) import template, which you can find here: https://s3-us-west-2.amazonaws.com/templates-7cffcf56-bd96-4b1c-b45b-a5b42f282e46/importtemplate.csv.

Fill out the template with your server and application data.

Upload your import file to an Amazon S3 bucket, and make a note of it's Object URL. Your import file must be in the CSV format.

Use the console or the StartImportTask command with the AWS CLI or one of the AWS SDKs to import the records from your file.

For more information, including step-by-step procedures, see Migration Hub Import in the AWS Application Discovery Service User Guide.

There are limits to the number of import tasks you can create (and delete) in an AWS account. For more information, see <a href="https://docs.aws.amazon.com/application-discovery/latest/userguide/adsservice_limits.html">AWS Application Discovery Service Limits in the AWS Application Discovery Service User Guide.

fn stop_continuous_export(
&self,
input: StopContinuousExportRequest
) -> RusotoFuture<StopContinuousExportResponse, StopContinuousExportError>

Stop the continuous flow of agent's discovered data into Amazon Athena.

fn stop_data_collection_by_agent_ids(
&self,
input: StopDataCollectionByAgentIdsRequest
) -> RusotoFuture<StopDataCollectionByAgentIdsResponse, StopDataCollectionByAgentIdsError>

Instructs the specified agents or connectors to stop collecting data.

fn update_application(
&self,
input: UpdateApplicationRequest
) -> RusotoFuture<UpdateApplicationResponse, UpdateApplicationError>

Updates metadata about an application.

Loading content...