Struct rusoto_discovery::DiscoveryClient

pub struct DiscoveryClient { /* fields omitted */ }

A client for the AWS Application Discovery Service API.

Methods

impl DiscoveryClient

pub fn new(region: Region) -> DiscoveryClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DiscoveryClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,