Trait rusoto_discovery::Discovery

pub trait Discovery {
    fn associate_configuration_items_to_application(
        &self, 
        input: AssociateConfigurationItemsToApplicationRequest
    ) -> RusotoFuture<AssociateConfigurationItemsToApplicationResponse, AssociateConfigurationItemsToApplicationError>;

    fn create_application(
        &self, 
        input: CreateApplicationRequest
    ) -> RusotoFuture<CreateApplicationResponse, CreateApplicationError>;

    fn create_tags(
        &self, 
        input: CreateTagsRequest
    ) -> RusotoFuture<CreateTagsResponse, CreateTagsError>;

    fn delete_applications(
        &self, 
        input: DeleteApplicationsRequest
    ) -> RusotoFuture<DeleteApplicationsResponse, DeleteApplicationsError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsRequest
    ) -> RusotoFuture<DeleteTagsResponse, DeleteTagsError>;

    fn describe_agents(
        &self, 
        input: DescribeAgentsRequest
    ) -> RusotoFuture<DescribeAgentsResponse, DescribeAgentsError>;

    fn describe_configurations(
        &self, 
        input: DescribeConfigurationsRequest
    ) -> RusotoFuture<DescribeConfigurationsResponse, DescribeConfigurationsError>;

    fn describe_export_configurations(
        &self, 
        input: DescribeExportConfigurationsRequest
    ) -> RusotoFuture<DescribeExportConfigurationsResponse, DescribeExportConfigurationsError>;

    fn describe_export_tasks(
        &self, 
        input: DescribeExportTasksRequest
    ) -> RusotoFuture<DescribeExportTasksResponse, DescribeExportTasksError>;

    fn describe_tags(
        &self, 
        input: DescribeTagsRequest
    ) -> RusotoFuture<DescribeTagsResponse, DescribeTagsError>;

    fn disassociate_configuration_items_from_application(
        &self, 
        input: DisassociateConfigurationItemsFromApplicationRequest
    ) -> RusotoFuture<DisassociateConfigurationItemsFromApplicationResponse, DisassociateConfigurationItemsFromApplicationError>;

    fn export_configurations(
        &self
    ) -> RusotoFuture<ExportConfigurationsResponse, ExportConfigurationsError>;

    fn get_discovery_summary(
        &self
    ) -> RusotoFuture<GetDiscoverySummaryResponse, GetDiscoverySummaryError>;

    fn list_configurations(
        &self, 
        input: ListConfigurationsRequest
    ) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>;

    fn list_server_neighbors(
        &self, 
        input: ListServerNeighborsRequest
    ) -> RusotoFuture<ListServerNeighborsResponse, ListServerNeighborsError>;

    fn start_data_collection_by_agent_ids(
        &self, 
        input: StartDataCollectionByAgentIdsRequest
    ) -> RusotoFuture<StartDataCollectionByAgentIdsResponse, StartDataCollectionByAgentIdsError>;

    fn start_export_task(
        &self, 
        input: StartExportTaskRequest
    ) -> RusotoFuture<StartExportTaskResponse, StartExportTaskError>;

    fn stop_data_collection_by_agent_ids(
        &self, 
        input: StopDataCollectionByAgentIdsRequest
    ) -> RusotoFuture<StopDataCollectionByAgentIdsResponse, StopDataCollectionByAgentIdsError>;

    fn update_application(
        &self, 
        input: UpdateApplicationRequest
    ) -> RusotoFuture<UpdateApplicationResponse, UpdateApplicationError>;
}

Trait representing the capabilities of the AWS Application Discovery Service API. AWS Application Discovery Service clients implement this trait.

Required Methods

fn associate_configuration_items_to_application(
&self,
input: AssociateConfigurationItemsToApplicationRequest
) -> RusotoFuture<AssociateConfigurationItemsToApplicationResponse, AssociateConfigurationItemsToApplicationError>

Associates one or more configuration items with an application.

fn create_application(
&self,
input: CreateApplicationRequest
) -> RusotoFuture<CreateApplicationResponse, CreateApplicationError>

Creates an application with the given name and description.

fn create_tags(
&self,
input: CreateTagsRequest
) -> RusotoFuture<CreateTagsResponse, CreateTagsError>

Creates one or more tags for configuration items. Tags are metadata that help you categorize IT assets. This API accepts a list of multiple configuration items.

fn delete_applications(
&self,
input: DeleteApplicationsRequest
) -> RusotoFuture<DeleteApplicationsResponse, DeleteApplicationsError>

Deletes a list of applications and their associations with configuration items.

fn delete_tags(
&self,
input: DeleteTagsRequest
) -> RusotoFuture<DeleteTagsResponse, DeleteTagsError>

Deletes the association between configuration items and one or more tags. This API accepts a list of multiple configuration items.

fn describe_agents(
&self,
input: DescribeAgentsRequest
) -> RusotoFuture<DescribeAgentsResponse, DescribeAgentsError>

Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an ID.

fn describe_configurations(
&self,
input: DescribeConfigurationsRequest
) -> RusotoFuture<DescribeConfigurationsResponse, DescribeConfigurationsError>

Retrieves attributes for a list of configuration item IDs. All of the supplied IDs must be for the same asset type (server, application, process, or connection). Output fields are specific to the asset type selected. For example, the output for a server configuration item includes a list of attributes about the server, such as host name, operating system, and number of network cards.

For a complete list of outputs for each asset type, see Using the DescribeConfigurations Action.

fn describe_export_configurations(
&self,
input: DescribeExportConfigurationsRequest
) -> RusotoFuture<DescribeExportConfigurationsResponse, DescribeExportConfigurationsError>

Deprecated. Use DescribeExportTasks instead.

Retrieves the status of a given export process. You can retrieve status from a maximum of 100 processes.

fn describe_export_tasks(
&self,
input: DescribeExportTasksRequest
) -> RusotoFuture<DescribeExportTasksResponse, DescribeExportTasksError>

Retrieve status of one or more export tasks. You can retrieve the status of up to 100 export tasks.

fn describe_tags(
&self,
input: DescribeTagsRequest
) -> RusotoFuture<DescribeTagsResponse, DescribeTagsError>

Retrieves a list of configuration items that are tagged with a specific tag. Or retrieves a list of all tags assigned to a specific configuration item.

fn disassociate_configuration_items_from_application(
&self,
input: DisassociateConfigurationItemsFromApplicationRequest
) -> RusotoFuture<DisassociateConfigurationItemsFromApplicationResponse, DisassociateConfigurationItemsFromApplicationError>

Disassociates one or more configuration items from an application.

fn export_configurations(
&self
) -> RusotoFuture<ExportConfigurationsResponse, ExportConfigurationsError>

Deprecated. Use StartExportTask instead.

Exports all discovered configuration data to an Amazon S3 bucket or an application that enables you to view and evaluate the data. Data includes tags and tag associations, processes, connections, servers, and system performance. This API returns an export ID that you can query using the DescribeExportConfigurations API. The system imposes a limit of two configuration exports in six hours.

fn get_discovery_summary(
&self
) -> RusotoFuture<GetDiscoverySummaryResponse, GetDiscoverySummaryError>

Retrieves a short summary of discovered assets.

fn list_configurations(
&self,
input: ListConfigurationsRequest
) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>

Retrieves a list of configuration items according to criteria that you specify in a filter. The filter criteria identifies the relationship requirements.

fn list_server_neighbors(
&self,
input: ListServerNeighborsRequest
) -> RusotoFuture<ListServerNeighborsResponse, ListServerNeighborsError>

Retrieves a list of servers that are one network hop away from a specified server.

fn start_data_collection_by_agent_ids(
&self,
input: StartDataCollectionByAgentIdsRequest
) -> RusotoFuture<StartDataCollectionByAgentIdsResponse, StartDataCollectionByAgentIdsError>

Instructs the specified agents or connectors to start collecting data.

fn start_export_task(
&self,
input: StartExportTaskRequest
) -> RusotoFuture<StartExportTaskResponse, StartExportTaskError>

Begins the export of discovered data to an S3 bucket.

If you specify agentIds in a filter, the task exports up to 72 hours of detailed data collected by the identified Application Discovery Agent, including network, process, and performance details. A time range for exported agent data may be set by using startTime and endTime. Export of detailed agent data is limited to five concurrently running exports.

If you do not include an agentIds filter, summary data is exported that includes both AWS Agentless Discovery Connector data and summary data from AWS Discovery Agents. Export of summary data is limited to two exports per day.

fn stop_data_collection_by_agent_ids(
&self,
input: StopDataCollectionByAgentIdsRequest
) -> RusotoFuture<StopDataCollectionByAgentIdsResponse, StopDataCollectionByAgentIdsError>

Instructs the specified agents or connectors to stop collecting data.

fn update_application(
&self,
input: UpdateApplicationRequest
) -> RusotoFuture<UpdateApplicationResponse, UpdateApplicationError>

Updates metadata about an application.